requirements.txt
streamlit
openai
anthropic
python-docx
markdown
beautifulsoup4
lxml
PyMuPDF
